QPager header
